# Conflicts:
#	src/main/resources/application-druid.yml

+++ src/main/java/com/aaa/project/system/userAccount/mapper/UserAccountMapper.java
@@ -24,7 +24,7 @@ public interface UserAccountMapper
 	 * @param userId
 	 * @return
 	 */
-	public UserAccount selectUserAccountByUserId(Integer userId);
+	public UserAccount selectUserAccountByUserId(Long userId);
 
 	/**
 	 * 查询洗车员所属店铺
